add: test save to db

+++ _finalProject/raise/src/main/java/com/ilyabuglakov/raise/dal/dao/interfaces/RoleDao.java
@@ -8,7 +8,7 @@ import java.sql.SQLException;
 import java.util.Set;
 
 public interface RoleDao extends Dao<Role> {
-    Set<String> getPermissions(Long id) throws DaoOperationException;
+    Set<String> getPermissions(Integer id) throws DaoOperationException;
     Set<String> getPermissions(String name) throws DaoOperationException, SQLException;
-    Set<String> getRoleNames(Long userId) throws SQLException, DaoOperationException;
+    Set<String> getRoleNames(Integer userId) throws SQLException, DaoOperationException;
 }
