+++ _finalProject/raise/src/main/java/com/ilyabuglakov/raise/dal/dao/interfaces/UserDao.java
@@ -11,5 +11,5 @@ import java.util.Optional;
 public interface UserDao extends Dao<User> {
     List<User> readAll() throws DaoOperationException;
     Optional<User> findByEmail(String email) throws DaoOperationException;
-    Optional<Long> getUserId(String email) throws SQLException, DaoOperationException;
+    Optional<Integer> getUserId(String email) throws SQLException, DaoOperationException;
 }
